refactor(codec): moving RegisterModuleConcrete method to utilities
