+++ go/documents/base/codec.go
@@ -6,7 +6,7 @@ package base
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
 
-	codecUtilities "github.com/AssetMantle/schema/go/codec"
+	codecUtilities "github.com/AssetMantle/schema/go/codec/utilities"
 )
 
 func RegisterLegacyAminoCodec(legacyAmino *codec.LegacyAmino) {
